dracut: create sub-directories for dracut modules
With the introduction of the Makefile it is now easy to move files into
different sub-directories. Thus create a sub-directory for each of the
dracut modules and move their files there. This not only cleans up the
main directory but also simplifies the Makefile and prevents bugs like
the one fixed in 5109f11 ("Makefile: Fix early-kdump file names").

One nice site effect by using 'cp' instead of 'install' is that file
permissions from the repo are preserved. So instead of manually setting
the file permissions for each call to 'install' we can now simply change
(and commit) the file permissions in the repo. Like it is done for
99kdumpbase/monitor_dd_progress.sh and 99zz-fadumpinit/module-setup.sh
in this commit.

Also adjust the .editorconfig to the new structure.

Signed-off-by: Philipp Rudo <prudo@redhat.com>