mkdumprd: add support for 95squash-erofs
With dracut 104 support for erofs in 99squash was added. For that the
squashfs specific code was split from 99squash module into
95squash-squashfs and a new 95squash-erofs was added. The modules are
structured the way, that you can either add 99squash, which then picks
the 'best' back end, or one of the 95squash-{squashfs,erofs} if you want
to make sure which back end is used.

Unfortunately erofs doesn't support the same compression algorithms
squashfs supports.  So explicitly set which image type we want so we can
set the correct --squash-compressor option.

Keep support for the old 99squash for the time being so newer versions
of kdump-utils can work with dracut <= 103.

Signed-off-by: Philipp Rudo <prudo@redhat.com>